+++ CHANGELOG.md
@@ -43,6 +43,8 @@ Main (unreleased)
 
 - The `remote.http` component can optionally define a request body. (@tpaschalis)
 
+- Added support for `loki.write` to flush WAL on agent shutdown. (@thepalbi)
+
 ### Bugfixes
 
 - Update `pyroscope.ebpf` to fix a logical bug causing to profile to many kthreads instead of regular processes https://github.com/grafana/pyroscope/pull/2778 (@korniltsev)